requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
openpyxl>=3.0.0
requests>=2.26.0
tabulate>=0.8.0
scipy>=1.7.0
python-dotenv>=0.19.0
pm4py>=2.2.0  # Process Mining for Python
pyinstaller>=5.7.0  # For creating executable 
